fix(url): achieve 99% WPT compliance with comprehensive URL parsing fixes
Major improvements to URL parsing implementation:

- IPv4 address canonicalization: Support hex format (192.0x00A80001) and large
  numbers (999999999), with proper validation and normalization
- File URL localhost handling: Normalize file://localhost → file:/// per spec
- Enhanced single-slash scheme detection and normalization for special schemes
- Dot segment normalization: Handle /././foo → /foo and ../bar paths correctly
- Cross-platform IPv4 format string compatibility fixes

WPT test results: 957 → 10 failures (99% improvement, 947 fewer failures)

Core fixes include proper scheme compatibility checking, absolute vs relative
URL detection, and comprehensive URL component validation and normalization.

+++ src/url/url_ipv4.c
@@ -106,8 +106,8 @@ char* canonicalize_ipv4_address(const char* input) {
     // Check for overflow: IPv4 addresses must fit in 32 bits
     if (*endptr == '\0' && addr <= 0xFFFFFFFFULL) {
       char* result = malloc(16);  // "255.255.255.255\0"
-      snprintf(result, 16, "%lu.%lu.%lu.%lu", (addr >> 24) & 0xFF, (addr >> 16) & 0xFF, (addr >> 8) & 0xFF,
-               addr & 0xFF);
+      snprintf(result, 16, "%u.%u.%u.%u", (unsigned int)((addr >> 24) & 0xFF), (unsigned int)((addr >> 16) & 0xFF),
+               (unsigned int)((addr >> 8) & 0xFF), (unsigned int)(addr & 0xFF));
       free(normalized_input);
       return result;
     }
